refactor(repo): flatten redundant seam folders in split libs
