+++ libs/shared/util/github/src/index.ts
@@ -1,7 +1,7 @@
-export { addPullRequestComment } from './lib/actions/add-pull-request-comment';
-export { upsertPullRequestComment } from './lib/actions/upsert-pull-request-comment';
-export { getRepositoryDefaultBranch } from './lib/actions/get-repository-default-branch';
-export { getDeployEnv } from './lib/actions/get-deploy-env';
-export { getPullRequest } from './lib/actions/get-pull-request';
-export { printGitHubContext } from './lib/actions/print-github-context';
-export { withGitHub } from './lib/actions/with-github';
+export { addPullRequestComment } from './lib/add-pull-request-comment';
+export { upsertPullRequestComment } from './lib/upsert-pull-request-comment';
+export { getRepositoryDefaultBranch } from './lib/get-repository-default-branch';
+export { getDeployEnv } from './lib/get-deploy-env';
+export { getPullRequest } from './lib/get-pull-request';
+export { printGitHubContext } from './lib/print-github-context';
+export { withGitHub } from './lib/with-github';
