+++ libs/shared/util/misc/src/lib/array-to-record.spec.ts
