+++ libs/shared/util/misc/src/lib/find-down.ts
