+++ libs/shared/util/misc/src/lib/get-package-version.ts
