+++ libs/shared/util/misc/src/lib/whoami.ts
