+++ libs/shared/util/release/src/lib/revert-package-json.ts
